requirements.txt
streamlit
pandas
numpy
google-cloud-firestore